fixed bug with posted time

+++ app/src/main/java/com/testapp/topredditnews/data/response/Post.kt
@@ -7,7 +7,7 @@ data class Post(
     @SerializedName("author")
     val author: String,
     @SerializedName("created_utc")
-    val createdUtc: Double,
+    val createdUtc: Long,
     @SerializedName("num_comments")
     val numComments: Int,
     @SerializedName("thumbnail")
